Extract unit tests into seperate files
