+++ src/validation/recursive_validator.rs
@@ -121,507 +121,3 @@ impl RecursiveValidator {
         }
     }
 }
-
-// Thank you ChatGPT for suggesting some test cases, much appreciated <3
-#[cfg(test)]
-mod tests {
-    mod structs {
-        use crate::{diagnostics::Diagnostic, test_utils::tests::parse_and_validate};
-
-        #[test]
-        fn pointers_should_not_be_considered_as_cycle() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT
-                b : B;
-            END_STRUCT END_TYPE
-
-            TYPE B : STRUCT
-                a : REF_TO A;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 0);
-        }
-
-        #[test]
-        fn one_cycle_abca() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT
-                b : B;
-            END_STRUCT END_TYPE
-
-            TYPE B : STRUCT
-                c : C;
-            END_STRUCT END_TYPE
-
-            TYPE C : STRUCT
-                a : A;
-                e : e;
-            END_STRUCT END_TYPE
-            
-            TYPE E : STRUCT
-                a_int: INT;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 1);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure(
-                    "A -> B -> C -> A",
-                    vec![(18..19).into(), (102..103).into(), (186..187).into(), (18..19).into()]
-                )
-            );
-        }
-
-        #[test]
-        fn one_cycle_self_a() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT
-                a : A;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 1);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure("A -> A", vec![(18..19).into(), (18..19).into()])
-            );
-        }
-
-        #[test]
-        fn one_cycle_multiple_self_a() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT
-                a1 : A;
-                a2 : A;
-                a3 : A;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 1);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure("A -> A", vec![(18..19).into(), (18..19).into()])
-            );
-        }
-
-        #[test]
-        fn one_cycle_aba() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT
-                b : B;
-            END_STRUCT END_TYPE
-
-            TYPE B : STRUCT
-                a : A;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 1);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure(
-                    "A -> B -> A",
-                    vec![(18..19).into(), (102..103).into(), (18..19).into()]
-                )
-            );
-        }
-
-        #[test]
-        fn one_cycle_bcb() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT
-                b : B;
-            END_STRUCT END_TYPE
-            
-            TYPE B : STRUCT
-                c : C;
-            END_STRUCT END_TYPE
-            
-            TYPE C : STRUCT
-                b : B;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 1);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure(
-                    "B -> C -> B",
-                    vec![(114..115).into(), (210..211).into(), (114..115).into()]
-                )
-            );
-        }
-
-        #[test]
-        fn one_cycle_with_multiple_identical_members_aba() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT 
-                b1 : B;
-                b2 : B;
-                b3 : B;
-            END_STRUCT END_TYPE
-
-            TYPE B : STRUCT 
-                a : A;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 1);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure(
-                    "A -> B -> A",
-                    vec![(18..19).into(), (152..153).into(), (18..19).into()]
-                )
-            );
-        }
-
-        #[test]
-        fn two_cycles_aa_and_aba() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT
-                a : A;
-                b : B;
-            END_STRUCT END_TYPE
-            
-            TYPE B : STRUCT
-                a : A;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 2);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure("A -> A", vec![(18..19).into(), (18..19).into()])
-            );
-            assert_eq!(
-                diagnostics[1],
-                Diagnostic::recursive_datastructure(
-                    "A -> B -> A",
-                    vec![(18..19).into(), (137..138).into(), (18..19).into()]
-                )
-            );
-        }
-
-        #[test]
-        fn two_cycles_branch_cc_and_cec() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT
-                b : B;
-            END_STRUCT END_TYPE
-            
-            TYPE B : STRUCT
-                c : C;
-            END_STRUCT END_TYPE
-            
-            TYPE C : STRUCT
-                c : C;
-                e : E;
-            END_STRUCT END_TYPE
-            
-            TYPE E : STRUCT
-                c : C;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 2);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure("C -> C", vec![(210..211).into(), (210..211).into()])
-            );
-            assert_eq!(
-                diagnostics[1],
-                Diagnostic::recursive_datastructure(
-                    "C -> E -> C",
-                    vec![(210..211).into(), (329..330).into(), (210..211).into()]
-                )
-            );
-        }
-
-        #[test]
-        fn two_cycles_with_branch() {
-            let diagnostics = parse_and_validate(
-                "
-            TYPE A : STRUCT
-                b : B;
-            END_STRUCT END_TYPE
-
-            TYPE B : STRUCT
-                c : C;
-            END_STRUCT END_TYPE
-
-            TYPE C : STRUCT
-                e : E;
-            END_STRUCT END_TYPE
-
-            TYPE E : STRUCT
-                f : F;
-            END_STRUCT END_TYPE
-
-            TYPE F : STRUCT
-                g : G;
-                b : B;
-            END_STRUCT END_TYPE
-
-            TYPE G : STRUCT
-                h : H;
-            END_STRUCT END_TYPE
-
-            TYPE H : STRUCT
-                i : I;
-            END_STRUCT END_TYPE
-
-            TYPE I : STRUCT
-                f : F;
-            END_STRUCT END_TYPE
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 2);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure(
-                    "F -> G -> H -> I -> F",
-                    vec![
-                        (354..355).into(),
-                        (461..462).into(),
-                        (545..546).into(),
-                        (629..630).into(),
-                        (354..355).into()
-                    ]
-                )
-            );
-            assert_eq!(
-                diagnostics[1],
-                Diagnostic::recursive_datastructure(
-                    "B -> C -> E -> F -> B",
-                    vec![
-                        (102..103).into(),
-                        (186..187).into(),
-                        (270..271).into(),
-                        (354..355).into(),
-                        (102..103).into()
-                    ]
-                )
-            );
-        }
-    }
-
-    mod function_blocks {
-        use crate::{diagnostics::Diagnostic, test_utils::tests::parse_and_validate};
-
-        #[test]
-        fn one_cycle_aba_var() {
-            let diagnostics = parse_and_validate(
-                "
-                FUNCTION_BLOCK A
-                    VAR
-                        b : B;
-                    END_VAR
-                END_FUNCTION_BLOCK
-
-
-                FUNCTION_BLOCK B
-                    VAR
-                        a : A;
-                    END_VAR
-                END_FUNCTION_BLOCK
-                ",
-            );
-
-            assert_eq!(diagnostics.len(), 1);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure(
-                    "A -> B -> A",
-                    vec![(32..33).into(), (185..186).into(), (32..33).into(),]
-                )
-            );
-        }
-
-        #[test]
-        fn one_cycle_aba_input() {
-            let diagnostics = parse_and_validate(
-                "
-                FUNCTION_BLOCK A
-                    VAR_INPUT
-                        b : B;
-                    END_VAR
-                END_FUNCTION_BLOCK
-
-
-                FUNCTION_BLOCK B
-                    VAR_INPUT
-                        a : A;
-                    END_VAR
-                END_FUNCTION_BLOCK
-                ",
-            );
-
-            assert_eq!(diagnostics.len(), 1);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure(
-                    "A -> B -> A",
-                    vec![(32..33).into(), (191..192).into(), (32..33).into()]
-                )
-            );
-        }
-
-        #[test]
-        fn one_cycle_aba_output() {
-            let diagnostics = parse_and_validate(
-                "
-                FUNCTION_BLOCK A
-                    VAR_OUTPUT
-                        b : B;
-                    END_VAR
-                END_FUNCTION_BLOCK
-
-
-                FUNCTION_BLOCK B
-                    VAR_OUTPUT
-                        a : A;
-                    END_VAR
-                END_FUNCTION_BLOCK
-                ",
-            );
-
-            assert_eq!(diagnostics.len(), 1);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure(
-                    "A -> B -> A",
-                    vec![(32..33).into(), (192..193).into(), (32..33).into()]
-                )
-            );
-        }
-
-        #[test]
-        fn one_cycle_aba_inout() {
-            let diagnostics = parse_and_validate(
-                "
-                FUNCTION_BLOCK A
-                    VAR_IN_OUT
-                        b : B;
-                    END_VAR
-                END_FUNCTION_BLOCK
-
-
-                FUNCTION_BLOCK B
-                    VAR_IN_OUT
-                        a : A;
-                    END_VAR
-                END_FUNCTION_BLOCK
-                ",
-            );
-
-            // No recursion because VAR_IN_OUT are pointers
-            assert_eq!(diagnostics.len(), 0);
-        }
-
-        #[test]
-        fn two_cycles_with_branch() {
-            let diagnostics = parse_and_validate(
-                "
-            FUNCTION_BLOCK A
-                VAR_INPUT
-                    b : B;
-                END_VAR
-            END_FUNCTION_BLOCK
-
-            FUNCTION_BLOCK B
-                VAR_INPUT
-                    c : C;
-                END_VAR
-            END_FUNCTION_BLOCK
-
-            FUNCTION_BLOCK C
-                VAR_INPUT
-                    e : E;
-                END_VAR
-            END_FUNCTION_BLOCK
-
-            FUNCTION_BLOCK E
-                VAR_INPUT
-                    f : F;
-                END_VAR
-            END_FUNCTION_BLOCK
-            
-            FUNCTION_BLOCK F
-                VAR_INPUT
-                    g : G;
-                    b : B;
-                END_VAR
-            END_FUNCTION_BLOCK
-            
-            FUNCTION_BLOCK G
-                VAR_INPUT
-                    h : H;
-                END_VAR
-            END_FUNCTION_BLOCK
-            
-            FUNCTION_BLOCK H
-                VAR_INPUT
-                    i : I;
-                END_VAR
-            END_FUNCTION_BLOCK
-
-            FUNCTION_BLOCK I
-                VAR_INPUT
-                    f : F;
-                END_VAR
-            END_FUNCTION_BLOCK
-            ",
-            );
-
-            assert_eq!(diagnostics.len(), 2);
-            assert_eq!(
-                diagnostics[0],
-                Diagnostic::recursive_datastructure(
-                    "F -> G -> H -> I -> F",
-                    vec![
-                        (592..593).into(),
-                        (769..770).into(),
-                        (919..920).into(),
-                        (1057..1058).into(),
-                        (592..593).into()
-                    ]
-                )
-            );
-            assert_eq!(
-                diagnostics[1],
-                Diagnostic::recursive_datastructure(
-                    "B -> C -> E -> F -> B",
-                    vec![
-                        (166..167).into(),
-                        (304..305).into(),
-                        (442..443).into(),
-                        (592..593).into(),
-                        (166..167).into()
-                    ]
-                )
-            );
-        }
-    }
-}
