+++ src/validation/tests.rs
@@ -5,6 +5,8 @@ mod duplicates_validation_test;
 mod generic_validation_tests;
 mod literals_validation_tests;
 mod pou_validation_tests;
+mod recursive_function_block_tests;
+mod recursive_structs_test;
 mod reference_resolve_tests;
 mod statement_validation_tests;
 mod variable_validation_tests;
